Update facilitator and tests
